Create README(Comments).md
In the src folder you should find the different java files i created.

All the pictures ive posted in the whatsapp group should be working 

if you run the file from the LoginPage a temp username and password has been set to admin and administration username and password has been set to admin1.

the login and register buttons should work as well as the back button on the register page.
all buttons on the left hand side of main menu and admin menu are working as well as the category buttons and close button (marked as a red X) when a panel opens up.

To help those with the backend ive labeled each and every button and textfield to make it easier to locate. If you are using Netbeans when you open the java files and under design you will see everything you can use this to locate buttons and textfields to see where exactly each button is and right click on the button then select events followed by actions performed to give the button an action. 

please do leave me a message if anything is confusing or for recomendations
I thank you you for your patience   
kind regard